optimize Power Of Two solution

+++ PowerOfTwo.kt
@@ -4,6 +4,6 @@ import kotlin.math.log2
 
 class PowerOfTwo {
     fun isPowerOfTwo(n: Int): Boolean {
-        return log2(n.toDouble()).toDouble()%1 <= 0.0000000000001
+        return log2(n.toDouble())%1 <= 0.0000000000001
     }
 }
